Big Changes - see description
Door Rework: You can now go up to a door and press "T" to interact with it an go through it. We have animations for the doors opening if someone would like to implement that.
Level 1 Finished - keeping level 1 simple, with just some parkour to get to the door
Keeping assets between scenes - I figured out multiple ways to do this, and the easiest will to be to not "destroy" the assets we need when moving to other scenes, will implement when with Sebastian
Added a bunch of Assets - I added that ghost you wanted Sebastian, as well as some other packs with a couple other flying enemies, and a bunch of other land ones we can use.
SLOPES INTERACTIONS - this was an absolute bitchslap but we needed it. We can now traverse along any incline pretty much, we do just slide a little bit so I need to add materials with more friction to counter it. This will make staircase implementation possible now for Level 2. My idea is to like progress through a castle. So level 1 will be a dungeon, 2 will be the castle, and 3 will be the boss fight outside to finish it.